# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartStock AI - POS & Inventory Management API for Kenyan retailers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ─── PURCHASE ORDERS ───────────────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales transactions
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: paymentMethod
          in: query
          schema:
            type: string
            enum: [cash, mpesa, bank]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleInput"
      responses:
        "201":
          description: Sale created, stock decremented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          $ref: "#/components/responses/NotFound"
  /sales/daily-summary:
    get:
      operationId: getDailySalesSummary
      tags: [sales]
      summary: Get today's till summary / reconciliation
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
  /sales/{id}/mpesa-status:
    get:
      operationId: getMpesaPaymentStatus
      tags: [sales]
      summary: Check M-Pesa STK push payment status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: M-Pesa payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpesaStatus"
  /sales/mpesa/callback:
    post:
      operationId: mpesaCallback
      tags: [sales]
      summary: Daraja M-Pesa STK push callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Callback processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ResultCode:
                    type: integer
                  ResultDesc:
                    type: string

  # ─── ANALYTICS ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    # ─── Categories ──────────────────────────────────────────────────────────
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateSupplierInput:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name

    # ─── Purchase Orders ─────────────────────────────────────────────────────
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        receiptNumber:
          type: string
        paymentMethod:
          type: string
          enum: [cash, mpesa, bank]
        subtotal:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        amountPaid:
          type: number
          format: float
        changeDue:
          type: number
          format: float
        mpesaRef:
          type: string
          nullable: true
        bankRef:
          type: string
          nullable: true
        mpesaStatus:
          type: string
          enum: [pending, confirmed, failed]
          nullable: true
        customerPhone:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
      required:
        - id
        - receiptNumber
        - paymentMethod
        - subtotal
        - discountAmount
        - totalAmount
        - amountPaid
        - changeDue
        - createdAt
        - items
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        costPrice:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
        lineTotal:
          type: number
          format: float
      required:
        - id
        - productId
        - productName
        - sku
        - quantity
        - unitPrice
        - costPrice
        - discountAmount
        - lineTotal
    CreateSaleInput:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [cash, mpesa, bank]
        amountPaid:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
          default: 0
        mpesaPhone:
          type: string
          nullable: true
        bankRef:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required:
        - paymentMethod
        - amountPaid
        - items
    SaleItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
          default: 0
      required:
        - productId
        - quantity
        - unitPrice
    DailySummary:
      type: object
      properties:
        date:
          type: string
          format: date
        totalSales:
          type: integer
        totalRevenue:
          type: number
          format: float
        totalCogs:
          type: number
          format: float
        grossProfit:
          type: number
          format: float
        cashTotal:
          type: number
          format: float
        mpesaTotal:
          type: number
          format: float
        bankTotal:
          type: number
          format: float
        avgTransactionValue:
          type: number
          format: float
      required:
        - date
        - totalSales
        - totalRevenue
        - totalCogs
        - grossProfit
        - cashTotal
        - mpesaTotal
        - bankTotal
        - avgTransactionValue
    MpesaStatus:
      type: object
      properties:
        saleId:
          type: integer
        mpesaRef:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, failed]
        message:
          type: string
          nullable: true
      required:
        - saleId
        - status

    # ─── Analytics ───────────────────────────────────────────────────────────
...